requirements.txt
Flask
python-openid
psycopg2
